Add Eclipse for model projects
Include `.classpath` and `.project` files for model packages in the
repository.

Those files are created by Obeo Designer when loading the projects.
Their existence make it easy to open the projects, because the projects
will be automatically discovered when opening the parent directory.

Those files should not change too often.