Add maven2Nature and maven2Builder to metamodel's `.projects`
Opening the metamodel projects with a non-Eclipse IDE (neovim with
https://github.com/mfussenegger/nvim-jdtls in this case) automatically
adds those entries to `.project`.

It does not seem to break anything in Obeo Designer, so this should be
safe to keep around. The convenience of committing this to the
repository is that those files will not be marked as modified all the
time.

The only noticeable change for Obeo Designer is that when opening those
projects, Obeo Designer will show a popup suggesting installing
maven2Nature. Installing it can safely be ignored.