Use maven-wrapper to compile backend
Includes the regular benefits of using maven-wrapper (does not require
a specific version of Maven to be installed on the host environment).

Additionally, helps nvim-jdtls start the eclipse.jdt.ls (LSP) at the
correct working directory. It will use the `backend` directory as the
root, instead of using the root of the repository.
See
https://github.com/mfussenegger/nvim-jdtls/blob/adeca37bd8ac86ec1782f1a9f270827e7a64dac3/lua/jdtls/setup.lua#L217-L220
for more information on computing the root directory.

The backend's `.nvimrc` that sets the root directory for nvim-jdtls is
no longer necessary.