feat(metamodel): add support for ApplicationDataPins
ApplicationDataPins are pins that appear on the diagram byt are not a
part of any UnitCall. They are the interface of the
ComputationApplicationRelease.

Rules related to connecting edges are very similar to the rules for
connecting UnitCalls' DataPins together:
* source must be a Provided pin
* source.incoming must be empty
* target must be a Required pin
* target.incoming must be empty

Due to the similarities, the diagram tools for:
* creating DataFlows
* reconnecting source and target of DataFlows
use Service's `isValidEdgeSource` and `isValidEdgeTarget`. Those methods
contain the logic for the predicates for those tools.

The icons of ApplicationDataPins are similar to DataPins' icons, but
have a black vertical bar on their left side.