refactor(metamodel): change names of application data pin tools
Change the names of diagram tools for creating diagram pin
(ApplicationDataPin). Required/Provided is misleading is this case,
since "Required" data pins are actually the "out" ones, and "Provided"
are actually "in" - seems reversed.

The new names specify "in"/"out" from the diagram's point of view.
Hopefully it makes it less confusing.