feat(metamodel): make data pins of UnitCalls readonly
Change the `pins` collection of `UnitCall`s to be readonly and not
changeable. This way it is harder (but not impossible) to
attach/detach/modify those pins.

This is because the `ComputedDataPin`s should be managed automatically
when setting the `unit` reference for `UnitCall`.

It works well in Sirius Desktop (the property is disabled).
In Sirius Web, the property is enabled, but when trying to change it,
there is an error in the console and a cryptic error message is shown.