feat(metamodel): mark `declared` reference on ComputedDataPins readonly
Marking the property as readonly makes it harder (but not impossible) to
change that reference by the user. It should be avoided, because
ComputedDataPins are managed automatically.

The property cannot be changed in Sirius Desktop's _Semantic_ tab, but
it can be changed in the _Main_ tab when viewing the properties of a
`ComputedDataPin`.

Marking the property as readonly does not affect Sirius Web.

The modification of this property could be further disallowed by making
the property not `changeable`. However, there are 2 problems with this
solution:

1. The code that manages creating `ComputedDataPin`s automatically would
   not be able to call `setDeclared`. It would have to use _Anonymous
   Inner Class_ to set the `declared` property
   (https://www.techiedelight.com/initialize-object-java/) or have a new
   method on the interface.

2. Sirius Desktop cannot load a model when the `declared` property is
   not changeable. There are errors that a `DeclaredDataPin` is an
   invalid value for `declared`.

Thus, the `declared` reference is only marked as readonly.