feat(metamodel): show dialog to pick unit release when creating call
Modify the tool for creating Unit Calls to show a dialog instead.
The dialog allows the user to fill in the information about the unit
call in a convenient way.

They can choose:

* the name of the unit call
* the ComputationUnitRelease to call (required)

The user can confirm creation only after choosing a valid
ComputationUnitRelease

Dialogs are not currently supported in Sirius Web. When trying to use
that tool, there is an error in the UI:

> Invalid input type, "InvokeNodeToolOnDiagramInput" has been received
> while "InvokeNodeToolOnDiagramInput" was expected

and server logs show:

> e.s.w.e.c.m.ModelOperationHandlerSwitch : The model operation
> DialogModelOperationImpl is not currently supported

Thus, there is a separate tool to only create the Unit Call without
asking for the ComputationUnitRelease. This tool can be used in Sirius
Web.