experiment(metamodel): add an External Java Action tool
Add an example External Java Action and a tool that calls it.
The action logs a message to the console to prove the it is working.

The action is called correctly in Sirius Desktop.
It is not called in Sirius Web. There in an error message in server
logs:
> .e.s.w.e.c.m.ModelOperationHandlerSwitch : The model operation
> ExternalJavaActionCallImpl is not currently supported

This commit is only here to leave a sign that using External Java Action
was tried and it did not work. It will be reverted.