experiment(metamodel): modify model during deserialization
Modify the `EditingContextSearchService` to modify the model during
deserialization. This allows injecting ComputationUnitReleases into the
model for convenience.

This way the user will be able to select a unit from a predefined list
rather than have to define the unit themselves.

Sirius Components and EMF in general operate on Resources,
ResourceLists, and EditingContexts that hold them. Documents are a
Sirius Web-only concept.

Looks like loaded EditingContexts are cached once they are loaded and
`findById` is not invoked after the context is loaded. This is
unfortunate, because this does not allow for adding/updating
`ComputationUnitRelease`s once the model is loaded.

Thus, there is a separate endpoint `/reset` in the `DocumentController`
that disposes an editing context, unloading it from memory and removing
it from the cache. The next time the editing context is requested (the
user opens the containing project), the `EditingContextSearchService`'s
`findById` is invoked again.

Resetting an EditingContext that is open in the UI will result in the
model being removed and requires refreshing the page or going to the
projects page and opening the project again.

When one thread sleeps because the model is being loaded, other threads
are suspended. This results in other editors being unresponsive to any
model changes. Model changes are being suspended until the thread stops
sleeping. Looks like loading models is single-threaded.

This mechanism lets us reload the auto-added `ComputationUnitRelease`s.

To complete the functionality, the following needs to be done:

1. Get actual `ComputationUnitRelease`s.

   This will probably be done by adding a new endpoint to which the main
   CAL backend will periodically push updates about all the available
   ComputationUnitReleases.

   Once they are provided, they can be used by the
   `EditingContextSearchService` to inject into the model.

   Since `ComputationUnitRelease`s will already be in memory via other
   means, injecting them into the model should be fast and will not
   block the thread.

2. Remove unused `ComputationUnitRelease`s before persisting the
   Document.

   This way the models saved to the database do not contain unused
   `ComputationUnitRelease`s and are simpler.