chore(nvim): add backend-specific nvimrc
Add an `.nvimrc` file that sets the root_dir for jdtls to the backend
directory. No matter which Java file is opened inside neovim, the same
LSP client will be used.

This makes it easier to work on the whole project and jump around files,
because all the files are loaded into memory.