chore(metamodel): commit maven nature for CAL tests project
The maven nature is added by eclipse.jdt.ls. It is not strictly needed,
but it is added automatically every time the language server loads the
project and leaves the git status dirty. Thus, it makes sense to commit
it once to have a cleaner list of changes in the project (git status).

Other CAL projects already have maven natures added in
6e5203b10b5576af048fe53522ac7074667bb125