refactor(ci): substitute secrets in settings.xml and .npmrc
Change the approach of passing secrets (GitHub token) to settings.xml
and .npmrc to allow pulling packages from GitHub Packages.

Instead of using environment variables directly when running commands,
generate settings.xml and .npmrc with the secrets substituted.

This is necessary to allow passing files with secrets to the build when
building a Docker image with BuildKit. BuildKit has a special mounting
type for files with secrets that does not leave any trace of the secrets
in the final image. Using environment variables and regular build
arguments would be risky.

https://docs.docker.com/develop/develop-images/build_enhancements/#new-docker-build-secret-information