chore(build): add a Dockerfile and a script to build the image
Add support for running the application in Docker.

The steps of the Dockerfile are very similar to the GitHub Workflow that
builds the application.

The Dockerfile uses Moby BuildKit to cache the Maven repository between
builds and to manage mounting secret files.

Resources:
* https://github.com/moby/buildkit/blob/master/frontend/dockerfile/docs/syntax.md
* https://docs.docker.com/develop/develop-images/build_enhancements/#new-docker-build-secret-information

Building and running the application with Docker is described in the
README.