fix(formatting): format sirius-web upstream changes with prettier
One file included in the patch from the upstream `sirius-web` repository
did not match prettier formatting. Since it was a patch that was applied
with `git am`, the pre-commit hook did not run for it.

The incorrect formatting causes CI to fail.

The commit that broke `main` is 73864e36e1daa23d6059f821929d6701fb1e89fb