experiment: get contents of the model in the frontend
Programmatically get contents of the entire model in the frontend.

There is no simple API to do it. Querying for diagram nodes returns only
objects that appear in the diagram, which for CAL are not all objects.

Subscribing to `treeEvent` returns all the objects in the model, but
incrementally - we need to simulate expanding nodes at each level,
essentially implementing Breadth First Search (BFS) where getting the
children of nodes requires a round-trip to the backend (to be concrete,
requires changing the `expanded` object IDs and waiting for the next
`treeEvent` to arrive with the children expanded).

The current solution only retrieves the structure of the model. Getting
object properties has to be done by subscribing to `propertiesEvent` and
is out of scope of this commit.

Closes #13