chore: add a hint about "sha1 information is lacking or useless"
This error can appear when trying to apply a patch and the upstream
repository is not a remote in the main repository. In this case, git
does not know about the parent commit hashes and cannot find a common
ancestor.

Executing the following commands solved the problem

git add remote upstream git@github.com:eclipse-sirius/sirius-web
git fetch upstream