feat(graphql): complete basic `createUnitCall` mutation
The mutation creates a `UnitCall` and connects it to a
`ComputationUnitRelease`, creating it if it does not exist.

There is a bug where the `ComputationUnitRelease` is always created. It
should be fixed soon.

Moreover, the UnitCall's pins appear outside of the node on the diagram
in the UI. This is not something that we have control over, sadly.