feat(frontend): automatically select the only representation
Expect that there will be just a single representation in the project
and select it automatically when the workbench loads.

In the final application the user will be editing just a single project,
and each project will have 1 model with a single representation.

This also allows the frontend to automatically find out what the
representation ID is. It will be required to pass to the workbench.