chore(frontend): copy the Workbench component from sirius components
The Workbench component is not customizable by consumers. There is no
easy way to inject custom components into the workbench UI. We will want
to add a BalticLSC toolbox UI next to the editor.

This commit copies the `Workbench` component (
https://github.com/eclipse-sirius/sirius-components/blob/master/frontend/src/workbench/Workbench.tsx
) along with any child components that it uses (like `Panels`,
`LeftSite`, `ValidationWebSocketContainer`, and `RightSite`). This gives
us full control over the UI.

Some props/components were not copied because they will not be needed in
the final application. For example the _Representations_ accordion in
the right panel was not copied, because there will be just a single
representation in the model.

Related to #18
See https://github.com/Gelio/CAL-web/issues/18#issuecomment-967889436