feat(ui): add toolbox for creating unit calls
Add a toolbox above the diagram displaying available units to call.
Clicking a unit will send a `createUnitCall` mutation to the Sirius Web
backend, which will in turn create a unit call and connect it to a
corresponding unit (use an existing one or create a new one if it does
not exist).

Units available to call are fetched from the BalticLSC backend,
similarly to how the BalticLSC diagram editor does it.

Authentication token is currently hardcoded. It will be customizable in
GH-43.

The root object ID is also hardcoded at the moment.

There are TODO comments that specify UI improvements that are left
before completing this task.

Related to #18