feat(validation): evaluate CAL semantic validation rules
Modify the ValidationDescriptionProvider to additionally run CAL
semantic validation rules and emit their diagnostics.

A CAL semantic validation rule receives a
`ComputationApplicationRelease` and returns a list of diagnostics for
that object. It can use EMF metamodel's methods and properties to come
up with those diagnostics.

There are no new semantic validation rules in this commit. They will be
added in future commits.

Closes #55