feat(thesis): pull EE thesis paper template
Use the thesis paper template document for the Faculty of Electrical
Engineering at the Warsaw University of Technology for the thesis text.

Source:
https://github.com/SP5LMA/EE-dyplom

Found at:
https://isod.ee.pw.edu.pl/isod-portal/dokumenty

The thesis template is pulled in as-is, without any noteworthy changes.
The only changes done to the source are ones done by the pre-commit hook
(fixing line endings, removing trailing whitespace, formatting the
README with prettier).