fix(ui): show onboard area when the project is empty
Right after creating the project, there are no models or representations
in the project. The automation for it is not complete. The user is
greeted with an unhelpful message about the representation being
missing.

This commit changes it so that the user is shown the `OnboardArea`
component instead and can create a model and a new representation for
it. The moment the user selects a representation, they are dropped into
our own `Workbench` with the toolbox.

The `OnboardArea` component is not exposed from sirius-components, so we
need to show the original `Workbench` component instead and switch the
components later. This is not great, because child comopnents lose state
(for example, the explorer tree is collapsed after selecting a
representation), but at least it works.

Closes #87