feat(ui): persist BalticLSC token in local storage
The BalticLSC authentication token is stored in `localStorage` and
retrieved during application load.

This happens at the `EditProjectView` page level. It means that when the
`Workbench` is embedded in #92, the authentication token can be
alternatively passed directly, without any effects from `localStorage`.
In other words, this functionality only applies when CAL-web is working
as a standalone application.

This commit also refactors how the toolbox authentication token controls
(the edit icon and the error icon) are injected into the toolbox. This
now happens via a prop on `Workbench`, which again allows for skipping
that part when using the `Workbench` directly in #92.

Closes #43