feat(auth): use BalticLSC JWT to determine the user
Refactor the server to expect a JWT to be passed in the `Authorization`
header with each API request. Then, use the subject field (`sub`) from
that JWT's body to determine the user. User accounts are created
on-demand.

The password for user account does not matter. The assumption here is
that JWTs will come from BalticLSC directly and Sirius Web will share
the same JWTs. Each user will have their own account with their own
projects. This way users do not modify each other's models.

This commit only changes the backend server. The frontend is not
modified yet. This means the frontend will not work correctly at the
moment, because no request contains the JWT.

Related to #96