feat(frontend): send Authorization header in API requests
The JWT is now sent in the `Authorization` header in HTTP API requests.

It is not handled for WebSockets, though, as custom headers are not sent
by browsers for WebSocket connections.
See https://stackoverflow.com/questions/4361173/http-headers-in-websockets-client-api/41521871

This makes all GraphQL subscriptions fail. The diagram editor heavily
relies on those subscriptions, which results in the application being
broken.

See https://github.com/Gelio/CAL-web/issues/96#issuecomment-980572837
for more information.