fix(auth): list only projects owned by the user
Restrict projects shown the user to only the projects they own.

This is only respected for GraphQL queries, because only in that context
do we know the actual username.

For GraphQL subscriptions, due to lack of support for WebSocket
authentication
(https://github.com/eclipse-sirius/sirius-components/issues/846),
there is no authentication.

Fixes #107