feat(model): add icons for required application data pins
Required application data pins have a black bar on the right side.
Provided ones keep the bar on the left.

This commit also refactors the style customizations to use a service and
a more predictable pattern of finding the right image. In the end, the
number of style customizations required for this feature is down from 8
to 1 and most of the complexity is handled in the metamodel's service
which concatenates the path for a single pin.

The same logic is used for changing the image for the computed data pin.

Due to a bug in Sirius Components, AQL does not work correctly in style
customizations. This means this change breaks images in Sirius Web.
See https://github.com/eclipse-sirius/sirius-components/issues/865

Related to #24