feat(auth): show auth token details in the modal
Display auth JWT parsing status in the model used to change it.
Show the username and when the token expires. Display parsing errors, if
any.

The expiry date is displayed relatively to now (e.g. "in 5 minutes").
There is a `Refresher` component that updates the UI every second to
make sure this time distance stays accurate even if the modal is shown
longer.

Closes #102
Closes #105