fix(thesis): do not refer to the root of the model
The model is technically a graph, not a tree, even though the term
"tree" is commonly used in Sirius Web and Sirius Desktop.

Let's at least not refer to the root of the model, because a root is
often associated with a tree. Let's use the term "the main model
object".