refactor(thesis): improve figures and listings arrangement
Change the order and positions of listings and figures to show them
closer to the text that refers to them.

Using `h` is not always the best idea. Sometimes just going with the `t`
or `b` position (maybe with a bang) improves the placement, as then the
figure breaks the paragraph, but appears closer to the text that refers
to it.

Sometimes moving the figure before the text that refers to it in LaTeX
results in a better layout than natural flow.