refactor(thesis): update thesis legal statements
Add the new legal statements (asserting authorship and giving WUT
license to use the thesis) in the front pages.

Use the document templates from the Electical Faculty website (ISOD).
They are not filled with any personal information.

It is expected to print the thesis out and then sign them, or print just
those 2 statements, sign them, scan them, and replace the PDFs in the
`gfx` directory to have a final thesis PDF with signed statements.

Closes #153
Closes #154